Fresh commit
I know this is a very bad practice to do many things in 1 commit. I read an
article about commits, and I will be careful and try to get used to following
a good habit. However, there were too many things that has been changed before
I read that article, so this commit is a mess.
Here is a list of things that has been changed in this commit.
* Add UML diagram
* Modify .gitignore to ignore Apache Derby dir, autosave file from Dia
* Add classes according to the UML diagram
* Add Tester that currently tests the Records class