Fix bug where tables are not recognized as existing even when they do
The cause of this was because hard coded lower-case names were used for creating tables,
but Derby automatically converted the table names to upper-case. Solutions to
this bug were to use upper-case table names and make sure that the table name
is upper-case by converting them intentionally in the constructor.