Add the four abstract methods inherited from Database
This commit implements the methods from the Database class. Only the insert
method is implemented. However, it might be a better choice to move these four
methods to the parent class because it seems that there is nothing
particularly different between the Records and Preferences
class, which are the only classes that the four methods are currently planned
on being implemented.