Implement getWidth method for the Word class
This method adds up the width of the Letters in the word. The width of the
word is added up to find the width of the line, and the width of the line
is used to determine whether the last word in the line should be moved to the
next line or not.