Fix bug where letter moves forward even when typing the incorrect key
This is caused because the (previously) getCurrentLetter method advances the
iterator regardless of the correctness of the keystroke. In order to fix this,
the method is called to advance the iterator only when the keystroke is
correct. The name of the method is changed to nextLetter, which is more
appropriate.