Add method that creates a LinkedHashMap of column names and values
This implementation aims the creation of LinkedHashMap required for select()
and update() methods in the Database class to become easier. However, issue
may remain regarding removing the "ID" column so that this column doesn't get
updated. In particular, the clone()-ing of the columnNamesAndTypes is not
tested thoroughly. There may be a need for another method that creates a list
of column names excluding the ID column(s). The implementation for the added
method is yet to be done for the Records class.